Move common test function to an include file
